requirements.txt
# 依赖包
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
pillow>=10.0.0
requests>=2.31.0
